+++ app/widgets/Chat/Chat.php
@@ -1369,7 +1369,9 @@ class Chat extends \Movim\Widget\Base
         }
 
         $msgkey = '<' . $message->jidfrom;
-        $msgkey .= ($message->type == 'groupchat') ? cleanupId($message->resource, true) : '';
+        $msgkey .= ($message->type == 'groupchat' && $message->resource != null)
+                    ? cleanupId($message->resource, true)
+                    : '';
         $msgkey .= '>' . substr($message->published, 11, 5);
 
         $counter = count($this->_wrapper[$date]);