+++ app/widgets/ChatActions/ChatActions.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Reported;
 use App\Url;
 
 use Moxl\Xec\Action\Message\Retract;
@@ -11,10 +12,36 @@ class ChatActions extends \Movim\Widget\Base
     /**
      * @brief Get a Drawer view of a contact
      */
-    public function ajaxGetContact($jid)
+    public function ajaxGetContact(string $jid)
     {
         $c = new ContactActions;
         $c->ajaxGetDrawer($jid);
+    }
+
+    /**
+     * @brief Block the contact
+     */
+    public function ajaxBlockContact(string $jid)
+    {
+        $r = Reported::firstOrCreate(['id' => $jid]);
+
+        $this->user->reported()->syncWithoutDetaching([$r->id]);
+        $this->user->refreshBlocked();
+
+        Toast::send($this->__('blocked.account_blocked'));
+
+        $this->rpc('Chats_ajaxClose', $jid, true);
+    }
+
+    /**
+     * @brief Unblock the contact
+     */
+    public function ajaxUnblockContact(string $jid)
+    {
+        $this->user->reported()->detach($jid);
+        $this->user->refreshBlocked();
+
+        Toast::send($this->__('blocked.account_unblocked'));
     }
 
     /**